#747 #730: Incluir cambio en el fichero principal SQL y evitar duplicado de cambio SQL; actualizar lista de tickets resueltos.
git-svn-id: https://opengnsys.es/svn/branches/version1.1@4925 a21b9725-9963-47de-94b9-378ad31fedc9

+++ admin/Database/ogAdmBD.sql
@@ -93,7 +93,7 @@ CREATE TABLE IF NOT EXISTS `aulas` (
   `grupoid` int(11) DEFAULT NULL,
   `ubicacion` varchar(255) DEFAULT NULL,
   `comentarios` text,
-  `puestos` tinyint(4) DEFAULT NULL,
+  `puestos` smallint DEFAULT NULL,
   `horaresevini` tinyint(4) DEFAULT NULL,
   `horaresevfin` tinyint(4) DEFAULT NULL,
   `modomul` tinyint(4) NOT NULL,
